修正三个编译错误
很神秘，一年多过去了，这三个地方编译不通过。
改完更疑惑了，为什么之前可以编译通过？

+++ nonEuclidGraphics/src/nonEuclideanEngine/WorldExample.cpp
@@ -232,7 +232,7 @@ tensorf333 nonEuc::WorldExample::Hyperbolic1::gamma(const vecf3& u)
 
 		t, 0.0f, 0.0f,
 		0.0f, t*sin(u[0])*sin(u[0]), 0.0f,
-		0.0f, 0.0f, -1/(sina*sina*(pow(u[2],3)+pow(u[2],5))),
+		0.0f, 0.0f, float(-1/(sina*sina*(pow(u[2],3)+pow(u[2],5)))),
 	};
 }
